env tutto apposto e non ci sono più invalid read
