+++ src/ft_tokenizer.c
@@ -7,22 +7,25 @@ t_var_count token_double_quote(char *token, char *line,
 
 	in_quotes = 1;
 	token[count.j++] = line[count.i];
-
 	count.i++;
 	while (line[count.i] != '\0')
 	{
 		if (line[count.i] == '"')
+		{
 			in_quotes = 0;
+			count.i++;
+		}
 		else if (in_quotes == 0 && line[count.i] == ' ')
 			break ;
 		else if (line[count.i] == '$')
 			count = ft_ex_dollar(line, token, env, count);
 		else
+		{
 			token[count.j++] = line[count.i];
-		count.i++;
+			count.i++;
+		}
 	}
-	token[count.j++] = '"';
-	token[count.j] = '\0';
+	token[count.j] = '"';
 	return count;
 }
 
@@ -33,7 +36,6 @@ t_var_count	token_single_quote(char *token, char *line,
 
 	in_quotes = 1;
 	token[count.j++] = line[count.i];
-
 	count.i++;
 	while (line[count.i] != '\0')
 	{
@@ -47,8 +49,7 @@ t_var_count	token_single_quote(char *token, char *line,
 			token[count.j++] = line[count.i];
 		count.i++;
 	}
-	token[count.j++] = '\'';
-	token[count.j] = '\0';
+	token[count.j] = '\'';
 	return count;
 }
 
@@ -88,19 +89,18 @@ char	**ft_tokenizer(char *input, t_env_var *env)
 	token_count = 0;
 	if (ft_check_quote(input) == -1)
 		return (ft_putendl_fd("minishell->Error: Unclosed quotes", 1), NULL);
-	tokens = (char **)ft_calloc(ft_count_token(input) + 1, sizeof (char *));
-	while (input[count.i] != '\0')
+	tokens = ft_calloc(ft_count_token(input) + 1, sizeof (char **));
+	tmp = ft_count_token(input);
+	while (token_count < tmp)
 	{
 		while (is_space(input[count.i]))
 			count.i++;
 		if (input[count.i] == '\0')
 			break;
-		tmp = count.i;
-		// printf("%d\n", ft_count(input, tmp, env));
-		tokens[token_count] = (char *)ft_calloc(ft_count(input, tmp, env), sizeof (char));
+		tokens[token_count] = (char *)ft_calloc(ft_count(input, count.i, env), sizeof (char));
 		count.j = 0;
-		count = token_separation(tokens[token_count++], input, env, count);
+		count = token_separation(tokens[token_count], input, env, count);
+		token_count++;
 	}
-	tokens[token_count] = NULL;
 	return (tokens);
 }
